Update Vertex to use one order connection per instrument

+++ xchange-stream-vertex/src/main/java/com/knowm/xchange/vertex/dto/Symbol.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 @ToString
 public class Symbol {
 
-  int product_id;
+  long product_id;
 
   String symbol;
